Changed SimpleQ to double because its default is 1e50, creating an overflow with precision-32.
--HG--
branch : week-of-code

+++ src/enzo/global_data.h
@@ -542,7 +542,7 @@ EXTERN int   AddParticleAttributes;
 EXTERN int   BigStarFormation;
 EXTERN int   BigStarFormationDone;
 EXTERN float BigStarSeparation;
-EXTERN float SimpleQ;
+EXTERN double SimpleQ;
 EXTERN float SimpleRampTime;
